# Large speed improvement for `inferences()`