--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -1,6 +1,6 @@
 Package: marginaleffects
 Title: Predictions, Comparisons, Slopes, Marginal Means, and Hypothesis Tests
-Version: 0.11.1.9012
+Version: 0.11.1.9013
 Authors@R: 
     c(person(given = "Vincent",
              family = "Arel-Bundock",