Related to #761. The bootstrap is slow because it calls `get_modeldata()` a lot inside `predictions()`, which is `FUN` in these lines:

https://github.com/vincentarelbundock/marginaleffects/blob/274d9195ecc48a3430de76562187c0ddbe89687c/R/bootstrap_rsample.R#L34-L44 

What I'm thinking is that you already know the model data, it's the line `modcall[["data"]] <- data`. So you could pass this data directly to `predictions()` and avoid running `get_modeldata()`. 

This leads to very large performance gains:

``` r
library(dplyr, warn.conflicts = FALSE)
library(marginaleffects)

set.seed(1234)
big_mtcars <- mtcars %>% 
  sample_n(15000, replace = TRUE)

mod <- lm(mpg ~ .,data=big_mtcars)
foo <- avg_predictions(mod) 

### Before
bench::mark(
  test = inferences(foo, method="rsample", R = 30)
)
#> Warning: Recommend at least 1000 non-missing bootstrap resamples for term `1`.
#> Recommend at least 1000 non-missing bootstrap resamples for term `1`.
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 test          33.2s    33.2s    0.0301    5.25GB     1.02

### After
bench::mark(
  test = inferences(foo, method="rsample", R = 30)
)
#> Warning: Recommend at least 1000 non-missing bootstrap resamples for term `1`.
#> Recommend at least 1000 non-missing bootstrap resamples for term `1`.
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 test          6.48s    6.48s     0.154     830MB     1.54
```

However, there are no numerical tests (only checks for class of `inferences()` output). I did a few quick checks but these are not comprehensive.

Also, I was limited in time so I just made this for `rsample` but it should be similar for the others (+ I'd like to be sure this is fine before I spend time on the rest).

It is still surprising that it takes several seconds for just 30 reps and 15,000 obs.