@etiennebacher 2023-05-09
Just for reference, updating this branch with #770 leads to even better changes:
``` r
bench::mark(
  test = inferences(foo, method="rsample", R = 30)
)
#> Warning: Recommend at least 1000 non-missing bootstrap resamples for term `1`.
#> Recommend at least 1000 non-missing bootstrap resamples for term `1`.
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 test          1.87s    1.87s     0.535     815MB     6.42
```