@etiennebacher 2023-05-10
I added a few numerical checks and the speedup for method "boot". For "fwb" I didn't see any speedup at all so I didn't change anything for this